Get higher order function for transaction wrapper working & move all business logic to service for homepage
