+++ package.json
@@ -20,6 +20,7 @@
 		"react": "16.13.1",
 		"react-dom": "16.13.1",
 		"react-select": "^3.1.0",
+		"serverless-mysql": "^1.5.4",
 		"sql-template-strings": "^2.2.2",
 		"styled-components": "5.0.1",
 		"swr": "^0.2.2"
